refactor models and it's features
