+++ street-goods-burger-backend/app/models/concerns/cart_validations.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+# CartValidations module
+module CartValidations
+  extend ActiveSupport::Concern
+  include ActiveModel::Validations
+
+  included do
+    belongs_to :store_transaction
+    has_many :food_orders
+
+    validates :quantity,
+              presence: true,
+              numericality: { only_integer: true, greater_than: -1 }
+    validates :total_price,
+              presence: true,
+              numericality: { greater_than: -0.99 }
+  end
+end
