+++ street-goods-burger-backend/app/models/food_order.rb
@@ -1,8 +1,6 @@
-class FoodOrder < ApplicationRecord
-  belongs_to :cart
+# frozen_string_literal: true
 
-  validates :food_name, length: { minimum: 1, maximum: 64 }, uniqueness: true
-  validates :food_category, presence: true, length: { minimum: 1, maximum: 64 }
-  validates :price, presence: true, numericality: { greater_than: -0.99 }
-  validates :quantity, presence: true, numericality: { only_integer: true, greater_than: 0 }
+# FoodOrder template
+class FoodOrder < ApplicationRecord
+  include FoodOrderValidations
 end
