+++ street-goods-burger-backend/app/models/store_admin.rb
@@ -1,19 +1,6 @@
+# frozen_string_literal: true
+
+# StoreAdmin template
 class StoreAdmin < ApplicationRecord
-  belongs_to :store
-
-  validates :email, presence: true,
-                    format: { with: /\A[\w+\-.]+@[a-z\d\-]+(\.[a-z]+)*\.[a-z]+\z/i }, # check email format email format
-                    uniqueness: true
-
-  validates :password, length: { minimum: 6 }
-
-  include Devise::JWT::RevocationStrategies::JTIMatcher
-  devise :database_authenticatable,
-         :registerable,
-         :validatable,
-         :recoverable,
-         :confirmable,
-         :rememberable,
-         :jwt_authenticatable,
-         jwt_revocation_strategy: self
+  include StoreAdminValidations
 end
